sm8350-common: Don't copy audio policy to qssi dirs
This depends on commit [1] which intends to pick a different audio
policy config on value-added system images and use to the regular
audio policy on gsi builds. However we do not provide the regular
config which leads to no usable audio policy config on GSI builds.
Our config does not rely on system value-added features, hence just
copy it to the regular path directly so that the same one will be
loaded on value-added and gsi builds.

[1]: https://github.com/LineageOS/android_system_media/commit/ce8ec8bff50d5c2ee48d6c63a235886f8715e495

Change-Id: I2ab532a4e8a120ed0c5495926dcce0a4b56d1f64